chore: Change directory structure
