+++ cmd/main.go
@@ -4,8 +4,8 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 	echoSwagger "github.com/swaggo/echo-swagger"
-	_ "koreatech-board-api/docs"
-	"koreatech-board-api/routes"
+	_ "koreatech-board-api/cmd/docs"
+	"koreatech-board-api/cmd/routes"
 	"os"
 )
 
@@ -16,7 +16,7 @@ import (
 //	@contact.name	Developer
 //	@contact.email	kongwoojin03@gmail.com
 
-//	@BasePath	/v3
+// @BasePath	/v3
 func main() {
 	// Echo instance
 	e := echo.New()
