+++ cmd/queries/cse.go
@@ -3,16 +3,16 @@ package queries
 import (
 	"fmt"
 	"github.com/labstack/echo/v4"
-	"koreatech-board-api/db"
-	"koreatech-board-api/model"
+	"koreatech-board-api/cmd/db"
+	"koreatech-board-api/cmd/model"
 	"math"
 	"net/http"
 	"strconv"
 )
 
 // @Summary		Get article list
-// @Description	Get ite article list
-// @Tags			ite
+// @Description	Get cse article list
+// @Tags			cse
 // @Accept			json
 // @Produce		json
 // @Param			board			path		string	true	"name of the board"
@@ -20,21 +20,26 @@ import (
 // @Param			num_of_items	query		integer	false	"items per page"
 // @Success		200				{object}	model.APIData
 // @Failure		404
-// @Router			/ite/{board} [get]
-func SelectIteQuery(c echo.Context) error {
+// @Router			/cse/{board} [get]
+func SelectCseQuery(c echo.Context) error {
 	boardRaw := c.Param("board")
 
 	var board = ""
 
 	switch boardRaw {
 	case "notice":
-		board = "247"
+		board = "notice"
+	case "free":
+		board = "freeboard"
+	case "job":
+		board = "jobboard"
+	case "pds":
+		board = "pds"
 	default:
 		return c.JSON(http.StatusNotFound, map[string]string{
 			"error": fmt.Sprintf("Board \"%s\" not found!", boardRaw),
 		})
 	}
-
 	page, pageErr := strconv.Atoi(c.QueryParam("page"))
 	numOfItems, noiErr := strconv.Atoi(c.QueryParam("num_of_items"))
 
@@ -52,7 +57,7 @@ func SelectIteQuery(c echo.Context) error {
 	listArgs := map[string]interface{}{"board": board, "offset": int64((page - 1) * numOfItems), "num_of_items": int64(numOfItems)}
 
 	var listQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT ite 
+		`SELECT cse 
 		{ id, num, title, writer, write_date, read_count }
 		FILTER .board = <str>$board order by contains(.num, '공지') DESC
 		THEN .write_date DESC
@@ -64,7 +69,7 @@ func SelectIteQuery(c echo.Context) error {
 	countArgs := map[string]interface{}{"board": board}
 
 	var countQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT count(ite filter .board=<str>$board)`,
+		`SELECT count(cse filter .board=<str>$board)`,
 		&count,
 		countArgs,
 	)
@@ -84,19 +89,19 @@ func SelectIteQuery(c echo.Context) error {
 }
 
 // @Summary		Get article
-// @Description	Get ite article by UUID
-// @Tags			ite
+// @Description	Get cse article by UUID
+// @Tags			cse
 // @Accept			json
 // @Produce		json
 // @Param			uuid	query		string	true	"uuid of article"
 // @Success		200		{object}	model.Article
 // @Failure		404
-// @Router			/article/ite [get]
-func IteArticleQuery(c echo.Context) error {
+// @Router			/article/cse [get]
+func CseArticleQuery(c echo.Context) error {
 	var results []model.Article
 
 	var articleQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT ite
+		`SELECT cse
 		{ id, title, writer, write_date, article_url, content, files: {file_name, file_url} }
 		FILTER .id = <uuid><str>$0`,
 		&results,
@@ -114,7 +119,7 @@ func IteArticleQuery(c echo.Context) error {
 
 // @Summary		Search article by title
 // @Description	Search article from specific board by title
-// @Tags			ite
+// @Tags			cse
 // @Accept			json
 // @Produce		json
 // @Param			board			path		string	true	"name of the board"
@@ -123,15 +128,21 @@ func IteArticleQuery(c echo.Context) error {
 // @Param			num_of_items	query		integer	false	"items per page"
 // @Success		200		{object}	model.Article
 // @Failure		404
-// @Router			/ite/{board}/search/title [get]
-func IteSearchWithTitleQuery(c echo.Context) error {
+// @Router			/cse/{board}/search/title [get]
+func CseSearchWithTitleQuery(c echo.Context) error {
 	boardRaw := c.Param("board")
 
 	var board = ""
 
 	switch boardRaw {
 	case "notice":
-		board = "247"
+		board = "notice"
+	case "free":
+		board = "freeboard"
+	case "job":
+		board = "jobboard"
+	case "pds":
+		board = "pds"
 	default:
 		return c.JSON(http.StatusNotFound, map[string]string{
 			"error": fmt.Sprintf("Board \"%s\" not found!", boardRaw),
@@ -156,7 +167,7 @@ func IteSearchWithTitleQuery(c echo.Context) error {
 	listArgs := map[string]interface{}{"board": board, "title": title, "offset": int64((page - 1) * numOfItems), "num_of_items": int64(numOfItems)}
 
 	var listQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT ite 
+		`SELECT cse 
 		{ id, num, title, writer, write_date, read_count }
 		FILTER .board = <str>$board and .title ilike <str>$title order by contains(.num, '공지') DESC
 		THEN .write_date DESC
@@ -168,7 +179,7 @@ func IteSearchWithTitleQuery(c echo.Context) error {
 	countArgs := map[string]interface{}{"board": board, "title": title}
 
 	var countQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT count(ite filter .board=<str>$board and .title ilike <str>$title)`,
+		`SELECT count(cse filter .board=<str>$board and .title ilike <str>$title)`,
 		&count,
 		countArgs,
 	)
