+++ cmd/queries/dorm.go
@@ -3,16 +3,16 @@ package queries
 import (
 	"fmt"
 	"github.com/labstack/echo/v4"
-	"koreatech-board-api/db"
-	"koreatech-board-api/model"
+	"koreatech-board-api/cmd/db"
+	"koreatech-board-api/cmd/model"
 	"math"
 	"net/http"
 	"strconv"
 )
 
 // @Summary		Get article list
-// @Description	Get arch article list
-// @Tags			arch
+// @Description	Get dorm article list
+// @Tags			dorm
 // @Accept			json
 // @Produce		json
 // @Param			board			path		string	true	"name of the board"
@@ -20,21 +20,19 @@ import (
 // @Param			num_of_items	query		integer	false	"items per page"
 // @Success		200				{object}	model.APIData
 // @Failure		404
-// @Router			/arch/{board} [get]
-func SelectArchQuery(c echo.Context) error {
+// @Router			/dorm/{board} [get]
+func SelectDormQuery(c echo.Context) error {
 	boardRaw := c.Param("board")
 
 	var board = ""
 
 	switch boardRaw {
 	case "notice":
-		board = "340"
+		board = "notice"
 	case "free":
-		board = "341"
+		board = "bulletin"
 	default:
-		return c.JSON(http.StatusNotFound, map[string]string{
-			"error": fmt.Sprintf("Board \"%s\" not found!", boardRaw),
-		})
+		return c.NoContent(http.StatusNotFound)
 	}
 
 	page, pageErr := strconv.Atoi(c.QueryParam("page"))
@@ -54,7 +52,7 @@ func SelectArchQuery(c echo.Context) error {
 	listArgs := map[string]interface{}{"board": board, "offset": int64((page - 1) * numOfItems), "num_of_items": int64(numOfItems)}
 
 	var listQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT arch 
+		`SELECT dorm 
 		{ id, num, title, writer, write_date, read_count }
 		FILTER .board = <str>$board order by contains(.num, '공지') DESC
 		THEN .write_date DESC
@@ -66,13 +64,12 @@ func SelectArchQuery(c echo.Context) error {
 	countArgs := map[string]interface{}{"board": board}
 
 	var countQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT count(arch filter .board=<str>$board)`,
+		`SELECT count(dorm filter .board=<str>$board)`,
 		&count,
 		countArgs,
 	)
 
 	if listQuery != nil || countQuery != nil {
-		fmt.Println(listQuery.Error())
 		return c.JSON(http.StatusBadRequest, map[string]string{
 			"error": "Query error!",
 		})
@@ -87,19 +84,19 @@ func SelectArchQuery(c echo.Context) error {
 }
 
 // @Summary		Get article
-// @Description	Get arch article by UUID
-// @Tags			arch
+// @Description	Get dorm article by UUID
+// @Tags			dorm
 // @Accept			json
 // @Produce		json
 // @Param			uuid	query		string	true	"uuid of article"
 // @Success		200		{object}	model.Article
 // @Failure		404
-// @Router			/article/arch [get]
-func ArchArticleQuery(c echo.Context) error {
+// @Router			/article/dorm [get]
+func DormArticleQuery(c echo.Context) error {
 	var results []model.Article
 
 	var articleQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT arch
+		`SELECT dorm
 		{ id, title, writer, write_date, article_url, content, files: {file_name, file_url} }
 		FILTER .id = <uuid><str>$0`,
 		&results,
@@ -107,7 +104,6 @@ func ArchArticleQuery(c echo.Context) error {
 	)
 
 	if articleQuery != nil {
-		fmt.Println(articleQuery.Error())
 		return c.JSON(http.StatusBadRequest, map[string]string{
 			"error": "Query error!",
 		})
@@ -118,7 +114,7 @@ func ArchArticleQuery(c echo.Context) error {
 
 // @Summary		Search article by title
 // @Description	Search article from specific board by title
-// @Tags			arch
+// @Tags			dorm
 // @Accept			json
 // @Produce		json
 // @Param			board			path		string	true	"name of the board"
@@ -127,17 +123,17 @@ func ArchArticleQuery(c echo.Context) error {
 // @Param			num_of_items	query		integer	false	"items per page"
 // @Success		200		{object}	model.Article
 // @Failure		404
-// @Router			/arch/{board}/search/title [get]
-func ArchSearchWithTitleQuery(c echo.Context) error {
+// @Router			/dorm/{board}/search/title [get]
+func DormSearchWithTitleQuery(c echo.Context) error {
 	boardRaw := c.Param("board")
 
 	var board = ""
 
 	switch boardRaw {
 	case "notice":
-		board = "340"
+		board = "notice"
 	case "free":
-		board = "341"
+		board = "bulletin"
 	default:
 		return c.JSON(http.StatusNotFound, map[string]string{
 			"error": fmt.Sprintf("Board \"%s\" not found!", boardRaw),
@@ -162,7 +158,7 @@ func ArchSearchWithTitleQuery(c echo.Context) error {
 	listArgs := map[string]interface{}{"board": board, "title": title, "offset": int64((page - 1) * numOfItems), "num_of_items": int64(numOfItems)}
 
 	var listQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT arch 
+		`SELECT dorm 
 		{ id, num, title, writer, write_date, read_count }
 		FILTER .board = <str>$board and .title ilike <str>$title order by contains(.num, '공지') DESC
 		THEN .write_date DESC
@@ -174,7 +170,7 @@ func ArchSearchWithTitleQuery(c echo.Context) error {
 	countArgs := map[string]interface{}{"board": board, "title": title}
 
 	var countQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT count(arch filter .board=<str>$board and .title ilike <str>$title)`,
+		`SELECT count(dorm filter .board=<str>$board and .title ilike <str>$title)`,
 		&count,
 		countArgs,
 	)
