+++ cmd/queries/ide.go
@@ -3,16 +3,16 @@ package queries
 import (
 	"fmt"
 	"github.com/labstack/echo/v4"
-	"koreatech-board-api/db"
-	"koreatech-board-api/model"
+	"koreatech-board-api/cmd/db"
+	"koreatech-board-api/cmd/model"
 	"math"
 	"net/http"
 	"strconv"
 )
 
 // @Summary		Get article list
-// @Description	Get dorm article list
-// @Tags			dorm
+// @Description	Get ide article list
+// @Tags			ide
 // @Accept			json
 // @Produce		json
 // @Param			board			path		string	true	"name of the board"
@@ -20,19 +20,21 @@ import (
 // @Param			num_of_items	query		integer	false	"items per page"
 // @Success		200				{object}	model.APIData
 // @Failure		404
-// @Router			/dorm/{board} [get]
-func SelectDormQuery(c echo.Context) error {
+// @Router			/ide/{board} [get]
+func SelectIdeQuery(c echo.Context) error {
 	boardRaw := c.Param("board")
 
 	var board = ""
 
 	switch boardRaw {
 	case "notice":
-		board = "notice"
+		board = "330"
 	case "free":
-		board = "bulletin"
+		board = "332"
 	default:
-		return c.NoContent(http.StatusNotFound)
+		return c.JSON(http.StatusNotFound, map[string]string{
+			"error": fmt.Sprintf("Board \"%s\" not found!", boardRaw),
+		})
 	}
 
 	page, pageErr := strconv.Atoi(c.QueryParam("page"))
@@ -52,7 +54,7 @@ func SelectDormQuery(c echo.Context) error {
 	listArgs := map[string]interface{}{"board": board, "offset": int64((page - 1) * numOfItems), "num_of_items": int64(numOfItems)}
 
 	var listQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT dorm 
+		`SELECT ide 
 		{ id, num, title, writer, write_date, read_count }
 		FILTER .board = <str>$board order by contains(.num, '공지') DESC
 		THEN .write_date DESC
@@ -64,7 +66,7 @@ func SelectDormQuery(c echo.Context) error {
 	countArgs := map[string]interface{}{"board": board}
 
 	var countQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT count(dorm filter .board=<str>$board)`,
+		`SELECT count(ide filter .board=<str>$board)`,
 		&count,
 		countArgs,
 	)
@@ -84,19 +86,19 @@ func SelectDormQuery(c echo.Context) error {
 }
 
 // @Summary		Get article
-// @Description	Get dorm article by UUID
-// @Tags			dorm
+// @Description	Get ide article by UUID
+// @Tags			ide
 // @Accept			json
 // @Produce		json
 // @Param			uuid	query		string	true	"uuid of article"
 // @Success		200		{object}	model.Article
 // @Failure		404
-// @Router			/article/dorm [get]
-func DormArticleQuery(c echo.Context) error {
+// @Router			/article/ide [get]
+func IdeArticleQuery(c echo.Context) error {
 	var results []model.Article
 
 	var articleQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT dorm
+		`SELECT ide
 		{ id, title, writer, write_date, article_url, content, files: {file_name, file_url} }
 		FILTER .id = <uuid><str>$0`,
 		&results,
@@ -114,7 +116,7 @@ func DormArticleQuery(c echo.Context) error {
 
 // @Summary		Search article by title
 // @Description	Search article from specific board by title
-// @Tags			dorm
+// @Tags			ide
 // @Accept			json
 // @Produce		json
 // @Param			board			path		string	true	"name of the board"
@@ -123,17 +125,17 @@ func DormArticleQuery(c echo.Context) error {
 // @Param			num_of_items	query		integer	false	"items per page"
 // @Success		200		{object}	model.Article
 // @Failure		404
-// @Router			/dorm/{board}/search/title [get]
-func DormSearchWithTitleQuery(c echo.Context) error {
+// @Router			/ide/{board}/search/title [get]
+func IdeSearchWithTitleQuery(c echo.Context) error {
 	boardRaw := c.Param("board")
 
 	var board = ""
 
 	switch boardRaw {
 	case "notice":
-		board = "notice"
+		board = "330"
 	case "free":
-		board = "bulletin"
+		board = "332"
 	default:
 		return c.JSON(http.StatusNotFound, map[string]string{
 			"error": fmt.Sprintf("Board \"%s\" not found!", boardRaw),
@@ -158,7 +160,7 @@ func DormSearchWithTitleQuery(c echo.Context) error {
 	listArgs := map[string]interface{}{"board": board, "title": title, "offset": int64((page - 1) * numOfItems), "num_of_items": int64(numOfItems)}
 
 	var listQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT dorm 
+		`SELECT ide 
 		{ id, num, title, writer, write_date, read_count }
 		FILTER .board = <str>$board and .title ilike <str>$title order by contains(.num, '공지') DESC
 		THEN .write_date DESC
@@ -170,7 +172,7 @@ func DormSearchWithTitleQuery(c echo.Context) error {
 	countArgs := map[string]interface{}{"board": board, "title": title}
 
 	var countQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT count(dorm filter .board=<str>$board and .title ilike <str>$title)`,
+		`SELECT count(ide filter .board=<str>$board and .title ilike <str>$title)`,
 		&count,
 		countArgs,
 	)
