+++ cmd/queries/ite.go
@@ -3,16 +3,16 @@ package queries
 import (
 	"fmt"
 	"github.com/labstack/echo/v4"
-	"koreatech-board-api/db"
-	"koreatech-board-api/model"
+	"koreatech-board-api/cmd/db"
+	"koreatech-board-api/cmd/model"
 	"math"
 	"net/http"
 	"strconv"
 )
 
 // @Summary		Get article list
-// @Description	Get emc article list
-// @Tags			emc
+// @Description	Get ite article list
+// @Tags			ite
 // @Accept			json
 // @Produce		json
 // @Param			board			path		string	true	"name of the board"
@@ -20,15 +20,15 @@ import (
 // @Param			num_of_items	query		integer	false	"items per page"
 // @Success		200				{object}	model.APIData
 // @Failure		404
-// @Router			/emc/{board} [get]
-func SelectEmcQuery(c echo.Context) error {
+// @Router			/ite/{board} [get]
+func SelectIteQuery(c echo.Context) error {
 	boardRaw := c.Param("board")
 
 	var board = ""
 
 	switch boardRaw {
 	case "notice":
-		board = "541"
+		board = "247"
 	default:
 		return c.JSON(http.StatusNotFound, map[string]string{
 			"error": fmt.Sprintf("Board \"%s\" not found!", boardRaw),
@@ -52,7 +52,7 @@ func SelectEmcQuery(c echo.Context) error {
 	listArgs := map[string]interface{}{"board": board, "offset": int64((page - 1) * numOfItems), "num_of_items": int64(numOfItems)}
 
 	var listQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT emc 
+		`SELECT ite 
 		{ id, num, title, writer, write_date, read_count }
 		FILTER .board = <str>$board order by contains(.num, '공지') DESC
 		THEN .write_date DESC
@@ -64,7 +64,7 @@ func SelectEmcQuery(c echo.Context) error {
 	countArgs := map[string]interface{}{"board": board}
 
 	var countQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT count(emc filter .board=<str>$board)`,
+		`SELECT count(ite filter .board=<str>$board)`,
 		&count,
 		countArgs,
 	)
@@ -84,19 +84,19 @@ func SelectEmcQuery(c echo.Context) error {
 }
 
 // @Summary		Get article
-// @Description	Get emc article by UUID
-// @Tags			emc
+// @Description	Get ite article by UUID
+// @Tags			ite
 // @Accept			json
 // @Produce		json
 // @Param			uuid	query		string	true	"uuid of article"
 // @Success		200		{object}	model.Article
 // @Failure		404
-// @Router			/article/emc [get]
-func EmcArticleQuery(c echo.Context) error {
+// @Router			/article/ite [get]
+func IteArticleQuery(c echo.Context) error {
 	var results []model.Article
 
 	var articleQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT emc
+		`SELECT ite
 		{ id, title, writer, write_date, article_url, content, files: {file_name, file_url} }
 		FILTER .id = <uuid><str>$0`,
 		&results,
@@ -114,7 +114,7 @@ func EmcArticleQuery(c echo.Context) error {
 
 // @Summary		Search article by title
 // @Description	Search article from specific board by title
-// @Tags			emc
+// @Tags			ite
 // @Accept			json
 // @Produce		json
 // @Param			board			path		string	true	"name of the board"
@@ -123,15 +123,15 @@ func EmcArticleQuery(c echo.Context) error {
 // @Param			num_of_items	query		integer	false	"items per page"
 // @Success		200		{object}	model.Article
 // @Failure		404
-// @Router			/emc/{board}/search/title [get]
-func EmcSearchWithTitleQuery(c echo.Context) error {
+// @Router			/ite/{board}/search/title [get]
+func IteSearchWithTitleQuery(c echo.Context) error {
 	boardRaw := c.Param("board")
 
 	var board = ""
 
 	switch boardRaw {
 	case "notice":
-		board = "541"
+		board = "247"
 	default:
 		return c.JSON(http.StatusNotFound, map[string]string{
 			"error": fmt.Sprintf("Board \"%s\" not found!", boardRaw),
@@ -156,7 +156,7 @@ func EmcSearchWithTitleQuery(c echo.Context) error {
 	listArgs := map[string]interface{}{"board": board, "title": title, "offset": int64((page - 1) * numOfItems), "num_of_items": int64(numOfItems)}
 
 	var listQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT emc 
+		`SELECT ite 
 		{ id, num, title, writer, write_date, read_count }
 		FILTER .board = <str>$board and .title ilike <str>$title order by contains(.num, '공지') DESC
 		THEN .write_date DESC
@@ -168,7 +168,7 @@ func EmcSearchWithTitleQuery(c echo.Context) error {
 	countArgs := map[string]interface{}{"board": board, "title": title}
 
 	var countQuery = db.Pool.Query(c.Request().Context(),
-		`SELECT count(emc filter .board=<str>$board and .title ilike <str>$title)`,
+		`SELECT count(ite filter .board=<str>$board and .title ilike <str>$title)`,
 		&count,
 		countArgs,
 	)
